change InferConfig
