+++ colossalai/inference/infer_struct.py
